refactor: modify how data gets collected and parsed
